Fixed callback execution multiple times.
The previous version of the code was only suitable in case when the callback or path/method had only one applicable middleware.

Setting multiple middleware for a one applicable controller resulted in the execution of callback the number of middleware applicable to that callback.

Now this can handle when any callback or path/method has over one middleware without repeating the callback multiple times.